+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - `pickStringRemember`: multi pick: if a picked value is an object its key-value pairs are put in the remember storage.
 ### Added
 - `pickStringRemember`: if a picked value is an object you can override the key used to pick a value from the remember storage: use `__key` in the object
+- `pickStringRemember`: property `joinByKey`: the picked value objects are joined by key using the separator into a single object.
 
 ## [1.65.6] 2024-10-07
 ### Added